roads update (#130)
* New post for roads updates

This represents the first roads update post Kelly Green

* Update 2016-04-18-utah-sgid-statewide-roads-data-layer-updates-4-15-2016.md

* Update 2016-04-18-utah-sgid-statewide-roads-data-layer-updates-4-15-2016.md

* Update 2016-04-18-utah-sgid-statewide-roads-data-layer-updates-4-15-2016.md

* Update 2016-04-18-utah-sgid-statewide-roads-data-layer-updates-4-15-2016.md

* Update 2016-04-18-utah-sgid-statewide-roads-data-layer-updates-4-15-2016.md

* updated text, could not create an ftp hyperlink

The ftp hyperlink issues seems to be a known one.  I just made some changes to the text to accommodate it.  I will try to incorporate more markdown in the next post for this subject due out in about 10-12 days

* Added one more thought

Told the users we will be adding dates next time