Documentation pillar review (#2680)
* chore: new monoquote for formatting citations

* chore: add archiving note

* chore: minor policy edits

* fix: missing summary section

* chore: minor discover fixes

* chore: missed on

* fix: sgid and sharing

* chore: main turn page minor fixes

* chore: one more turn minor fix

* wip: turn connection checkpoint

* feat: updated TURN docs

* refactor: add parent p tags to mono and default spacing

* refactor: convert i to em

In earlier versions of the HTML specification, the <i> element was merely a presentational element used to display text in italics, much like the <b> element was used to display text in bold letters. This is no longer true, as these tags now define semantics rather than typographic appearance. A browser will typically still display the contents of the <i> element in italic type, but is, by definition, no longer required to do so. To display text in italic type, authors should use the CSS font-style property.

* chore: grammar?

* refactor: b -> strong

---------

Co-authored-by: steveoh <sgourley@utah.gov>