Update 2024-12-13-statewide-lidar-products-now-available.mdx
a couple misc. edits

I think this looking great!

The only problem I have is we have covered as much of the state as we can but really do not have "statewide coverage" considering the hole over dugway proving grounds. The hole is in the graphic on page https://gis.utah.gov/products/sgid/elevation/lidar/#2022-western--eastern-utah
@miriamseely can you make some adjustments to accommodate this fact?